Make the JournalLayout have explicit methods to add newline/hspace
Previously each item that was inserted in the layout was supposed to say
whether it wanted newlines around it.

Now, we have a cleaner, more reusable layout model where you can insert
the following into a JournalLayout:

  - appendItem (item) - Assumes that the item has an item.actor - this is
    what is normally used to insert visible items like events and headings.

  - appendHSpace () - Puts a horizontal space after the last item that got
    appended.  The space comes from the item-spacing CSS attribute.

  - appendNewline () - Puts a newline after the last item that got inserted,
    and of course moves the "cursor" to the first column.

So, now it is the responsibility of the caller to form the layout in terms
of items, spaces, and newlines - like a word processor.

In the future we may have items that grow horizontally and such, and we may
need to add the concept of boxes and glue, like in TeX.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>